created 1-array_iterator.c file

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,17 +1,17 @@
 #include "function_pointers.h"
 
 /**
- *
- *
- *
- *
+ * array_iteration - A function that sprints all element of a array
+ * @array: An input to the function
+ * @size: An input to the function
+ * @action: An input to the function
  */
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
 	if (!array || !action)
 		return;
-	int i = 0;
+	size_t i = 0;
 
 	for (; i < size; i++)
 		action(array[i]);
